+++ react/__mocks__/vtex.styleguide/Modal.js
@@ -1,0 +1,3 @@
+import Modal from '@vtex/styleguide/lib/Modal'
+
+export default Modal