+++ react/components/ProductAssemblyContext/Group.tsx
@@ -45,10 +45,7 @@ export const ProductAssemblyGroupContext = createContext<
   AssemblyOptionGroupState | undefined
 >(undefined)
 
-export const ProductAssemblyGroupContextProvider: FC<ProductAssemblyGroupContextProviderProps> = ({
-  assemblyOption,
-  children,
-}) => {
+const initState = (assemblyOption: AssemblyOptionGroupState) => {
   const groupPath = getGroupPath(assemblyOption.treePath)
   const quantitySum = Object.values(assemblyOption.items ?? {}).reduce(
     (acc, { quantity }) => acc + quantity,
@@ -62,15 +59,20 @@ export const ProductAssemblyGroupContextProvider: FC<ProductAssemblyGroupContext
     return acc
   }, {})
 
-  const initialState: AssemblyOptionGroupState = {
-    ...assemblyOption,
-    path: groupPath,
-    quantitySum,
-    optin: assemblyOption.required,
-    valuesOfInputValues,
-  }
+  assemblyOption.path = assemblyOption.path ?? groupPath
+  assemblyOption.quantitySum = assemblyOption.quantitySum ?? quantitySum
+  assemblyOption.optin = assemblyOption.optin ?? assemblyOption.required
+  assemblyOption.valuesOfInputValues =
+    assemblyOption.valuesOfInputValues ?? valuesOfInputValues
 
-  const [state, dispatch] = useReducer(reducer, initialState)
+  return assemblyOption
+}
+
+export const ProductAssemblyGroupContextProvider: FC<ProductAssemblyGroupContextProviderProps> = ({
+  assemblyOption,
+  children,
+}) => {
+  const [state, dispatch] = useReducer(reducer, assemblyOption, initState)
 
   return (
     <ProductAssemblyDispatchContext.Provider value={dispatch}>
@@ -107,6 +109,19 @@ export const useProductAssemblyGroupDispatch = () =>
 export const useProductAssemblyGroupState = () =>
   useContext(ProductAssemblyGroupContext)
 
+const checkSamePath = (path1: string[], path2: string[]) => {
+  return path1.join('/') === path2.join('/')
+}
+
+const getGroupState = (
+  state: AssemblyOptionGroupState,
+  groupPath: string[]
+) => {
+  return checkSamePath(state.path, groupPath)
+    ? (state as AssemblyOptionGroupState)
+    : (path(groupPath, state) as AssemblyOptionGroupState)
+}
+
 function reducer(
   state: AssemblyOptionGroupState,
   action: DispatchAction
@@ -114,7 +129,7 @@ function reducer(
   switch (action.type) {
     case 'OPTIN': {
       const { groupPath } = action.args
-      const groupState = path(groupPath, state) as AssemblyOptionGroupState
+      const groupState = getGroupState(state, groupPath)
 
       groupState.optin = !groupState.optin
 
@@ -122,12 +137,9 @@ function reducer(
     }
     case 'SET_INPUT_VALUE': {
       const { groupPath, inputValue, inputValueLabel } = action.args
-      const groupState = path(groupPath, state) as AssemblyOptionGroupState
+      const groupState = getGroupState(state, groupPath)
 
-      groupState.valuesOfInputValues = {
-        ...groupState.valuesOfInputValues,
-        [inputValueLabel]: inputValue,
-      }
+      groupState.valuesOfInputValues[inputValueLabel] = inputValue
 
       return { ...state }
     }