+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Make Input Values work for recursive assemblies.
+
 ## [2.7.0] - 2020-03-23
 
 ### Added